renamed nvmlmetrics pkg to nvidia
