+++ pkg/collector/corechecks/gpu/nvidia/collector.go
@@ -5,12 +5,12 @@
 
 //go:build linux
 
-// Package nvmlmetrics holds the logic to collect metrics from the NVIDIA Management Library (NVML).
+// Package nvidia holds the logic to collect metrics from the NVIDIA Management Library (NVML).
 // The main entry point is the BuildCollectors functions, which returns a set of collectors that will
 // gather metrics from the available NVIDIA devices on the system. Each collector is responsible for
 // a specific subsystem of metrics, such as device metrics, GPM metrics, etc. The collected metrics will
 // be returned with the associated tags for each device.
-package nvmlmetrics
+package nvidia
 
 import (
 	"errors"
