+++ pkg/collector/corechecks/gpu/nvidia/collector_test.go
@@ -5,7 +5,7 @@
 
 //go:build linux
 
-package nvmlmetrics
+package nvidia
 
 import (
 	"errors"
