+++ pkg/collector/corechecks/gpu/nvidia/convert.go
@@ -5,7 +5,7 @@
 
 //go:build linux
 
-package nvmlmetrics
+package nvidia
 
 import (
 	"bytes"
